Add secrets (#3)
* remove dubble jsonencoding and adding rolling updates for envs

* checksums tested

* done
