+++ Tiltfile
@@ -1,8 +1,8 @@
-k8s_yaml("k8s/signup-d.yaml")
-docker_build("signup", "services/signup", live_update=[sync("services/signup/src", "/app/src")])
+k8s_yaml("k8s/signup-d.yaml") # this line tells tilt to apply the k8s/signup-d.yaml file and to monitor it (simply put, it tells tilt 'this file is under your responsibility')
+docker_build("signup", "services/signup", live_update=[sync("services/signup/src", "/app/src")]) # this line tells tilt to build the docker image of the signup service and to monitor the src folder for changes (basically, 'build it, rebuild in case of changes, specifically in the src folder just inject the changes to avoid unnecessary rebuild for every code change')
 # make sure debugging port is equal to the launch file of the application (9001 in this case)
 # make sure to update the port is wright when using post-man (3001 in this case) 
-k8s_resource('signup-d', port_forwards=['3001:3000','9001:9229']) 
+k8s_resource('signup-d', port_forwards=['3001:3000','9001:9229']) # this line tells tilt to create a port-forward from the machine to the containers (used for: attaching the debugger, testing the service, connecting to rabbitmq if the services run outside the cluster, etc.)
 
 k8s_yaml("k8s/auth-d.yaml")
 docker_build("auth", "services/auth", live_update=[sync("services/auth/src", "/app/src")])
