+++ k8s/auth-d.yaml
@@ -11,6 +11,8 @@ spec:
     metadata:
       labels:
         app: auth
+      annotations:
+        secret_checksum: "uNGK7dxxkqjAlx6h712yMMYBIeySOQBB+Uwt9bK2y54="
     spec:
       containers:
       - name: auth
@@ -18,4 +20,6 @@ spec:
         ports:
         - containerPort: 3000
         - containerPort: 9229
-
+        envFrom:
+        - secretRef:
+            name: auth-secret
