+++ k8s/call-d.yaml
@@ -11,6 +11,8 @@ spec:
     metadata:
       labels:
         app: call
+      annotations:
+        secret_checksum: "NnOCIe4Dzk2Ml5RBXzBQQakegjmqlKygEYPpZkNqJ6s="
     spec:
       containers:
       - name: call
@@ -18,4 +20,6 @@ spec:
         ports:
         - containerPort: 3000
         - containerPort: 9229
-
+        envFrom:
+        - secretRef:
+            name: call-secret
