+++ k8s/chat-d.yaml
@@ -11,6 +11,8 @@ spec:
     metadata:
       labels:
         app: chat
+      annotations:
+        secret_checksum: "JeI7/aZAX4rorXwLLfRu5XF5y0TMe1EkX8tVAxs1x2c="
     spec:
       containers:
       - name: chat
@@ -18,4 +20,6 @@ spec:
         ports:
         - containerPort: 3000
         - containerPort: 9229
-
+        envFrom:
+        - secretRef:
+            name: chat-secret
