+++ k8s/notify-d.yaml
@@ -11,6 +11,8 @@ spec:
     metadata:
       labels:
         app: notify
+      annotations:
+        secret_checksum: "5RGatWyACTDAA4ZO9ygtpYqEfsAUc9V8DCSrC/1rfgk="
     spec:
       containers:
       - name: notify
@@ -18,4 +20,6 @@ spec:
         ports:
         - containerPort: 3000
         - containerPort: 9229
-
+        envFrom:
+        - secretRef:
+            name: notify-secret
