--- k8s/payments-d.yaml
@@ -1,21 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: payment-d
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: payment
-  template:
-    metadata:
-      labels:
-        app: payment
-    spec:
-      containers:
-      - name: payment
-        image: payment
-        ports:
-        - containerPort: 3000
-        - containerPort: 9229
-
