--- k8s/rating-d.yaml
@@ -1,21 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: review-d
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: review
-  template:
-    metadata:
-      labels:
-        app: review
-    spec:
-      containers:
-      - name: review
-        image: review
-        ports:
-        - containerPort: 3000
-        - containerPort: 9229
-
