+++ k8s/review-d.yaml
@@ -11,6 +11,8 @@ spec:
     metadata:
       labels:
         app: review
+      annotations:
+        secret_checksum: "hSsCu3KqqWuj7/4/lEOP+pNGA+PktoNUphWnfqPgyFQ="
     spec:
       containers:
       - name: review
@@ -18,4 +20,6 @@ spec:
         ports:
         - containerPort: 3000
         - containerPort: 9229
-
+        envFrom:
+        - secretRef:
+            name: review-secret
