+++ k8s/send-d.yaml
@@ -11,6 +11,8 @@ spec:
     metadata:
       labels:
         app: send
+      annotations:
+        secret_checksum: "p3kwXACr8yDnzMVPEei213EBP88x1bZzUtV18WRtS8w="
     spec:
       containers:
       - name: send
@@ -18,4 +20,6 @@ spec:
         ports:
         - containerPort: 3000
         - containerPort: 9229
-
+        envFrom:
+        - secretRef:
+            name: send-secret
