--- k8s/sokcets-d.yaml
@@ -1,21 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: socket-d
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: socket
-  template:
-    metadata:
-      labels:
-        app: socket
-    spec:
-      containers:
-      - name: socket
-        image: socket
-        ports:
-        - containerPort: 3000
-        - containerPort: 9229
-
