+++ main.tf
@@ -99,12 +99,17 @@ resource "kubernetes_ingress_v1" "ingress_rules" {
 
 # (Module) deployments of services under development
 module "deployments_of_our_services" {
-  source = "./modules/deployments-of-our-services"
+  source           = "./modules/deployments-of-our-services"
+  secret_checksums = module.secrets.secret_checksums
 }
 
 # (Module) databases
 module "dbs" {
   source = "./modules/dbs"
+}
+
+module "secrets" {
+  source = "./modules/secrets"
 }
 
 //todo: export to a model
